+++ src/LazyImage/CHANGELOG.md
@@ -7,4 +7,6 @@
     for more details.
 -   The `data-hd-src` attribute was changed to use a Stimulus value called `src`. See the
     updated README for usage.
+-   For both JavaScript events - `lazy-image:connect` and `lazy-image:ready` -
+    the `event.detail.hd` `Image` instance was moved to `event.detail.image`.
 -   Support added for Symfony 6